// src/users/users.enum.ts
export enum ERole {
  Admin = 'admin',
  Member = 'member',
  Teacher = 'teacher',
}

export enum EGender {
  male = 'male',
  female = 'female',
  unknown = 'unknown',
}

export enum ERegisterTeacher {
  pending = 'pending',
  reject = 'reject',
  notRegistered = 'notRegistered',
  accept = 'accept',
}